Fixed issue affecting WIC slot and PCMCIA controller registers
It was discovered that the original mod made to enable pin 31 of the ROM sockets to be pulled high to prevent the WE/ pin of ROMs smaller than 8mbit from being asserted was also resulting in an address input to the two CPLDs to also be pulled high.

The signal is A19, and corresponds with a bit mask of 0x00080000. It was only pure coincidence that this didnt affect the System Option Register or LED Control Register, because both of these live at addresses 0x0D08000X, so A19 would have been high to access these registers anyway.

Accesses to the WIC slot and PCMCIA controller registers, being at lower addresses, would therefore have resulted in accesses to invalid addresses, 0x0D0D00XX and 0x0D0B000X respectively.

More cut traces and a couple of mod wires were required to resolve this issue, further complicating the re-use of these routers.